+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix issue [#27](https://github.com/intersystems/language-server/issues/27): DeprecationWarning in LS Output channel at startup
 - Fix issue [#44](https://github.com/intersystems/language-server/issues/44): prompt for "use dark theme" if i'm using a dark vscode background
 - Fix issue [#48](https://github.com/intersystems/language-server/issues/48): Add .gitignore for OS files
+- Fix issue [#50](https://github.com/intersystems/language-server/issues/50): Go to definition for method of current class in client-side editing mode shouldn't open server version
 
 ## [1.0.3] - 2020-10-26
 - Fix issue [#41](https://github.com/intersystems/language-server/issues/41): Still scope for reducing number of web sessions 1.0.2 creates